requirements.txt
langchain==0.2.9
langchainhub==0.1.20
langchain-community==0.2.6
langchain_core==0.2.21
langchain_openai==0.1.17
langchain_text_splitters==0.2.2
openai==1.35.15
SQLAlchemy==2.0.31
streamlit==1.36.0
duckduckgo-search==6.2.1
pypdf==4.3.0
fastembed
docarray==0.40.0
validators==0.33.0
python-dotenv
faiss-cpu
yt_dlp
pydub
librosa
bs4
transformers
torch
streamlit-float
openai-whisper
streamlit-mic-recorder
clarifai
eva-decord==0.6.1
salesforce-lavis==1.0.2